미션1 PR 코드 리뷰 후 전체적인 리팩토링
- System.out.print() 대신에 Logger 사용

- index() 를 QuestionController 클래스로 옮김 questionList()의 static 사용을 막기위해

- 메서드 네이밍 잘못된거 고침

- 로직 중복 줄임
