+++ src/main/java/com/codessquad/qna/QuestionController.java
@@ -1,10 +1,13 @@
 package com.codessquad.qna;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
@@ -15,13 +18,27 @@ import java.util.List;
 @Controller
 public class QuestionController {
 
-    private static List<Question> questionList = new ArrayList<>();
-    public static List<Question> questionList() {
+    private final Logger LOGGER = LoggerFactory.getLogger(QuestionController.class.getName());
+
+    private List<Question> questionList = new ArrayList<>();
+
+    private List<Question> questionList() {
         return Collections.unmodifiableList(questionList);
     }
 
+    @GetMapping("/")
+    public String index(Model model) {
+        model.addAttribute("questionList", questionList());
+        return "index";
+    }
+
+    @GetMapping("/qna/form")
+    public String qnaForm() {
+        return "/qna/form";
+    }
+
     @PostMapping("/qna/create")
-    public String question(Question question, Model model) {
+    public String createQuestion(Question question, Model model) {
         String date = LocalDateTime.now().format(DateTimeFormatter.ofPattern("yyyy-MM-dd HH:mm"));
 
         question.setDate(date);
@@ -29,7 +46,7 @@ public class QuestionController {
         questionList.add(question);
         model.addAttribute("questionList", questionList);
 
-        System.out.println(question.toString());
+        LOGGER.info(question.toString());
         return "redirect:/";
     }
